# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: BangMusic Server API
  version: 1.0.0
  summary: 자가호스팅 BangMusic 음악 서버와 모바일 앱 사이의 계약
  description: |
    설계 초안(2026-10-02). 구현 전 계약이며 동작하는 서버가 아직 없다.

    공통 규칙
    - 모든 경로는 `/v1` 아래에 있다. 호환되지 않는 변경은 `/v2`로만 낸다.
    - `/v1` 안에서는 추가만 한다(새 필드·새 엔드포인트·새 enum 값). 앱은 모르는 필드와 enum 값을 무시해야 한다.
    - 서버가 무엇을 지원하는지는 `GET /v1/server`의 `api.minor`와 `features`로 판단한다.
    - 오류는 RFC 9457 `application/problem+json`이며 기계 판독용 `code`를 포함한다.
    - 응답에는 서버의 파일 경로가 절대 포함되지 않는다. 모든 자원은 불투명 ID로만 가리킨다.
    - 시간은 RFC 3339 UTC 문자열, 길이는 밀리초 정수다.
    - 권한이 없는 자원은 존재 여부를 숨기기 위해 404 `not_found`로 응답한다. 403은 역할이 부족한 관리 API에만 쓴다.
    - 모든 응답에 `X-Request-Id`가 있다. 앱은 오류 보고에 이 값을 표시한다.
    - 모든 동작은 `426 client_too_old`(앱 API minor가 서버 최소값보다 낮음)와 `503`(`setup_required`, `maintenance` 등)을 돌려줄 수 있다.
  license:
    name: 미정 (서버 패키지 라이선스는 제품 소유자 결정 사항)
    identifier: LicenseRef-TBD
servers:
  - url: https://{host}/v1
    description: 구매자가 운영하는 서버. 앱은 사용자가 입력한 주소를 사용한다.
    variables:
      host:
        default: music.example.net
...
security:
  - bearerAuth: []
paths:
  /server:
    get:
      tags: [server]
      operationId: getServerInfo
      summary: 서버 식별자·API 버전·기능 목록
      description: |
        앱이 서버 주소를 등록할 때와 앱 시작 시마다 호출한다. 인증 없이 호출 가능하며
        라이브러리 내용이나 사용자 정보는 포함하지 않는다.
      security: []
      responses:
        '200':
          description: 서버 정보
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServerInfo' }
              example:
                server_id: srv_01JB2M4X7Q9ZK3V5T8N6R1WY0E
                name: 우리집 음악
                product: bangmusic-server
                version: 1.0.0
                api: { major: 1, minor: 0 }
                min_client_api_minor: 0
                features: [transcode, lyrics, playlist_covers, history, change_feed]
                transcode_profiles: [aac_256, aac_128]
                limits: { max_page_size: 200, max_history_batch: 500, max_cover_bytes: 5242880 }
                setup_required: false
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
...
  /playlists/{playlistId}/cover:
    parameters:
      - { $ref: '#/components/parameters/PlaylistId' }
    put:
      tags: [playlists]
      operationId: putPlaylistCover
      summary: 플레이리스트 표지 업로드
      description: 서버가 디코딩 후 다시 인코딩하고 메타데이터(EXIF)를 제거한다.
      parameters:
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        '200':
          description: 저장됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Playlist' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '413':
          description: '`payload_too_large` — `limits.max_cover_bytes` 초과'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '415':
          description: '`unsupported_media_type`'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
    delete:
      tags: [playlists]
      operationId: deletePlaylistCover
      summary: 표지 제거 (곡 표지 모자이크로 돌아감)
      responses:
        '204': { description: 제거됨 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '426': { $ref: '#/components/responses/ClientTooOld' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <access_token>`. 액세스 토큰은 불투명 문자열이며 서버가 해시로 조회한다.'
    mediaTicket:
      type: apiKey
      in: query
      name: mt
      description: |
        렌디션·표지 요청 전용 단기 서명 티켓. 사용자·세션·대상 자원·만료 시각에 묶인다.
        다른 API에는 쓸 수 없다. 서버와 프록시는 접근 로그에서 `mt` 값을 가린다.
  parameters:
...
    PlaylistId:
      name: playlistId
      in: path
      required: true
      schema: { type: string }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 마지막으로 받은 ETag. 없으면 428.
      schema: { type: string }
...
  responses:
    BadRequest:
      description: '`validation_failed` — `errors`에 필드별 사유'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: |
        `access_expired`(갱신 후 재시도), `access_invalid`, `session_revoked`(재로그인 필요),
        `account_disabled`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example: { type: 'about:blank', title: Unauthorized, status: 401, code: access_expired, detail: 액세스 토큰이 만료되었습니다., request_id: req_9Kd1 }
...
    NotFound:
      description: '`not_found` — 없거나 접근 권한이 없음'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    PreconditionFailed:
      description: '`version_conflict` — 다른 기기가 먼저 수정함. 응답 ETag가 현재 버전.'
      headers:
        ETag: { schema: { type: string } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ClientTooOld:
      description: '`client_too_old` — 앱의 API 버전이 서버의 `min_client_api_minor`보다 낮음. 앱 업데이트 안내.'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: '`rate_limited` — `Retry-After` 초 뒤 재시도'
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unavailable:
      description: |
        `storage_unavailable`(음악 저장소 마운트 장애), `maintenance`(마이그레이션·복원 중),
        `setup_required`(관리자 계정 미생성). 지수 백오프로 재시도 가능.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: 기계 판독용 오류 코드. 전체 목록은 02-api-guide.md §오류 코드.
        detail: { type: string, description: 사람이 읽는 설명. 경로·SQL·스택을 포함하지 않는다. }
        request_id: { type: string }
        retryable: { type: boolean }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    ServerInfo:
      type: object
      required: [server_id, name, product, version, api, min_client_api_minor, features, limits, setup_required]
      properties:
        server_id: { type: string, description: 서버 설치의 영속 식별자. 앱의 서버 프로필·캐시 분리 키. }
        name: { type: string }
        product: { type: string, const: bangmusic-server }
        version: { type: string, description: 서버 소프트웨어 버전(SemVer) }
        api:
          type: object
          required: [major, minor]
          properties:
            major: { type: integer }
            minor: { type: integer, description: '`/v1` 안에서 기능이 추가될 때마다 증가' }
        min_client_api_minor: { type: integer }
        features:
          type: array
          description: 서버가 켜 둔 선택 기능. 앱은 모르는 값을 무시한다.
          items: { type: string }
        transcode_profiles:
          type: array
          items: { type: string }
        limits:
          type: object
          additionalProperties: { type: integer }
        setup_required: { type: boolean }
...
    Playlist:
      type: object
      required: [id, name, version, item_count, duration_ms, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        artwork_id: { type: [string, 'null'], description: 업로드한 표지. null이면 앱이 `mosaic_artwork_ids`로 모자이크를 만든다. }
        mosaic_artwork_ids: { type: array, maxItems: 4, items: { type: string } }
        version: { type: integer }
        item_count: { type: integer }
        duration_ms: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...